ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12

# STT
faster-whisper==1.0.3

# TTS — Piper (Nepali fallback, always loaded)
torch==2.4.1
torchaudio==2.4.1
scipy==1.14.1
numpy==1.26.4
piper-tts==1.4.2

# TTS — Coqui XTTSv2 (primary; ~1.8 GB model downloaded on first run)
# Model cache dir is controlled by TTS_HOME env var (default ~/.local/share/tts)
# Mount a Docker volume at that path to persist across container restarts.
TTS>=0.22.0

# Audio
soundfile==0.12.1

# Logging
structlog==24.4.0

# Runtime dependency pulled by faster-whisper utils
requests==2.32.3
